Drop the authentication code
* Authentication is now handled by Cockpit.
