+++ web/src/FirstUser.test.jsx
@@ -23,7 +23,7 @@ import React from "react";
 
 import { screen, waitFor, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { authRender } from "./test-utils";
+import { installerRender } from "./test-utils";
 import { createClient } from "./client";
 import FirstUser from "./FirstUser";
 
@@ -53,7 +53,7 @@ beforeEach(() => {
 });
 
 it("allows defining a new user", async () => {
-  authRender(<FirstUser />);
+  installerRender(<FirstUser />);
   const firstUser = await screen.findByText(/A user/);
   const button = within(firstUser).getByRole("button", { name: "is not defined" });
   userEvent.click(button);
@@ -86,7 +86,7 @@ it("allows defining a new user", async () => {
 });
 
 it("does not change anything if the user cancels", async () => {
-  authRender(<FirstUser />);
+  installerRender(<FirstUser />);
   const firstUser = await screen.findByText(/A user/);
   const button = within(firstUser).getByRole("button", { name: "is not defined" });
   userEvent.click(button);
@@ -112,7 +112,7 @@ describe("when the first user is already defined", () => {
   });
 
   it("allows removing the user", async () => {
-    authRender(<FirstUser />);
+    installerRender(<FirstUser />);
     const button = await screen.findByRole("button", { name: "jdoe" });
     userEvent.click(button);
