+++ web/src/test-utils.js
@@ -23,7 +23,6 @@ import React from "react";
 import { render } from "@testing-library/react";
 
 import { InstallerClientProvider } from "./context/installer";
-import { AuthProvider } from "./context/auth";
 import { createClient } from "./client";
 
 const InstallerProvider = ({ children }) => {
@@ -31,20 +30,8 @@ const InstallerProvider = ({ children }) => {
   return <InstallerClientProvider client={client}>{children}</InstallerClientProvider>;
 };
 
-const AllProviders = ({ children }) => {
-  return (
-    <InstallerProvider>
-      <AuthProvider>{children}</AuthProvider>
-    </InstallerProvider>
-  );
-};
-
 const installerRender = (ui, options = {}) => {
   return render(ui, { wrapper: InstallerProvider, ...options });
 };
 
-const authRender = (ui, options = {}) => {
-  return render(ui, { wrapper: AllProviders, ...options });
-};
-
-export { installerRender, authRender };
+export { installerRender };
